Develop/networking (#1)
* Networking Api added

* Buffer and Package tests added

* Event packages added

* Main GameApplication components made shared_ptr

* Ring buffer added

* Sending Events from Client to server implemented and tested

* tests runner added

* Tests hanging in release fixed

* Entity creation package added

* Phisics sync with network added

* Remove entity package added

* Multiple clients support added

* Rests reorganized + clientserver integration test added

* Unknown package type bug fixed

* random added

* server crash after disconnect fixed

* client disconnected signal added

* Mem network layer added

* Cmake reorganized

* Vector template added & events package optimized

* Docs added

* Linux Build fixed

* Windows build fixed

* Some warnings removed

* workflows tests fixed on windows

* workflows tests fixed on windows

* some warnings removed

+++ src/time/elapsedtimer.cpp
@@ -35,7 +35,7 @@ void ElapsedTimer::reset() {
 }
 
 double ElapsedTimer::progress() const {
-    const double milliseconds = Time::currentMilliseconds();
+    const double milliseconds = static_cast<double>(Time::currentMilliseconds());
     if(m_interval == 0)\
         return 0;
     return (milliseconds - m_checkPoint) / m_interval;
